+++ frontend/modules/build_course/controllers/VideoController.php
@@ -3,6 +3,8 @@
 namespace frontend\modules\build_course\controllers;
 
 use common\models\vk\Course;
+use common\models\vk\CourseNode;
+use common\models\vk\searchs\VideoFavoriteSearch;
 use common\models\vk\searchs\VideoSearch;
 use common\models\vk\TagRef;
 use common\models\vk\Tags;
@@ -65,11 +67,21 @@ class VideoController extends Controller
         $dataProvider = new ArrayDataProvider([
             'allModels' => array_values($result['data']['video']),
         ]);
+        
+        if(\Yii::$app->request->isAjax){
+            Yii::$app->getResponse()->format = 'json';
+            return [
+                'code'=> 200,
+                'page' => $result['filter']['page'],
+                'data' => array_values($result['data']['video']),
+                'message' => '请求成功！',
+            ];
+        }
       
         return $this->render('index', [
             'searchModel' => $searchModel,
             'filters' => $result['filter'],
-            'pagers' => $result['pager'],
+            'totalCount' => $result['total'],
             'dataProvider' => $dataProvider,
             'courseMap' => $this->getCourseByCreatedBy(),
         ]);
@@ -110,6 +122,10 @@ class VideoController extends Controller
         ]);
         $model->loadDefaultValues();
         
+        if(!ActionUtils::getInstance()->getIsHasEditNodePermission($model->courseNode->course_id)){
+            throw new NotFoundHttpException(Yii::t('app', 'You have no permissions to perform this operation.'));
+        }
+        
         if ($model->load(Yii::$app->request->post())) {
             Yii::$app->getResponse()->format = 'json';
             $result = ActionUtils::getInstance()->CreateVideo($model, Yii::$app->request->post());
@@ -126,7 +142,7 @@ class VideoController extends Controller
             return $this->renderAjax('create', [
                 'model' => $model,
                 'allRef' => $this->getVideoByReference(),
-                'allTeacher' => Teacher::getTeacherByLevel(Yii::$app->user->identity->customer_id),
+                'allTeacher' => Teacher::getTeacherByLevel(Yii::$app->user->id, 0, false),
                 'videoFiles' => Video::getUploadfileByVideo(),
                 'allTags' => ArrayHelper::map(Tags::find()->all(), 'id', 'name'),
             ]);
@@ -146,6 +162,10 @@ class VideoController extends Controller
     {
         $model = $this->findModel($id);
         
+        if(!ActionUtils::getInstance()->getIsHasEditNodePermission($model->courseNode->course_id)){
+            throw new NotFoundHttpException(Yii::t('app', 'You have no permissions to perform this operation.'));
+        }
+        
         if ($model->load(Yii::$app->request->post())) {
             Yii::$app->getResponse()->format = 'json';
             $result = ActionUtils::getInstance()->UpdateVideo($model, Yii::$app->request->post());
@@ -162,7 +182,7 @@ class VideoController extends Controller
             return $this->renderAjax('update', [
                 'model' => $model,
                 'allRef' => $this->getVideoByReference(),
-                'allTeacher' => Teacher::getTeacherByLevel($model->customer_id),
+                'allTeacher' => Teacher::getTeacherByLevel($model->created_by, 0, false),
                 'videoFiles' => Video::getUploadfileByVideo($model->source_id),
                 'allTags' => ArrayHelper::map(Tags::find()->all(), 'id', 'name'),
                 'tagsSelected' => array_keys(TagRef::getTagsByObjectId($id, 2)),
@@ -180,6 +200,10 @@ class VideoController extends Controller
     {
         $model = $this->findModel($id);
         
+        if(!ActionUtils::getInstance()->getIsHasEditNodePermission($model->courseNode->course_id)){
+            throw new NotFoundHttpException(Yii::t('app', 'You have no permissions to perform this operation.'));
+        }
+        
         if ($model->load(Yii::$app->request->post())) {
             Yii::$app->getResponse()->format = 'json';
             $result = ActionUtils::getInstance()->DeleteVideo($model);
@@ -195,13 +219,28 @@ class VideoController extends Controller
     }
     
     /**
-     * 引用 现有的 Video 模型。
+     * 引用 已有的 Video 模型。
      * 如果是post传值，返回成功的json数据，否则返回失败
-     * @param string $id
      * @return json
      */
-    public function actionReference($id)
+    public function actionReference()
     {
+        $searchModel = new VideoFavoriteSearch();
+        $result = $searchModel->referenceSearch(Yii::$app->request->queryParams);
+        
+        $dataProvider = new ArrayDataProvider([
+            'allModels' => array_values($result['data']['video']),
+        ]);
+        
+        return $this->renderAjax('reference', [
+            'searchModel' => $searchModel,
+            'filters' => $result['filter'],
+            'totalCount' => $result['total'],
+            'dataProvider' => $dataProvider,
+        ]);
+        
+        
+        
         $result = $this->findVideoByCiteInfo($id);
         Yii::$app->getResponse()->format = 'json';
         
@@ -264,9 +303,15 @@ class VideoController extends Controller
      */
     protected function getCourseByCreatedBy()
     {
-        $courses = Course::find()
-            ->where(['created_by' => Yii::$app->user->id])
-            ->all();
+        //根据已存在的视频查询课程id
+        $courseIds = Video::find()->select(['CourseNode.course_id'])
+            ->from(['Video' => Video::tableName()])
+            ->leftJoin(['CourseNode' => CourseNode::tableName()], '(CourseNode.id = Video.node_id AND CourseNode.is_del = 0)')
+            ->where(['Video.created_by' => Yii::$app->user->id, 'Video.is_del' => 0]);
+        
+        //查询课程
+        $courses = Course::find()->where(['created_by' => Yii::$app->user->id])
+            ->andWhere(['id' => $courseIds])->all();
         
         return ArrayHelper::map($courses, 'id', 'name');
     }
